use DI container and move config related stuff to Config service
I looked at different plugins from Shopware and they all use
this structure, so I adapted it.
A big plugin you can look at is SwagBundle.
Install it via the Shopware Store as a trial version and then copy it
out of your Shopware installation to get the source code.