only drop triggers if we have to
In plugin version < 2.0.0 the plugin added DB triggers to log changed articles.
In 2.0.0 we dropped those triggers completely. Under some circumstances the DB user is not allowed to create or drop triggers. Therefor the update script failed, even if there where no triggers installed.

This can happen if MySQL's binlog is active and the DB user that Shopware is using does not have SUPER privileges.

On installation of plugin version < 2.0.0 we catched the exception that was thrown when the DB user was not able to install the triggers and disabled the delta export. This should have been done in the update script for 2.0.0 as well. We forgot this use case 🙈

In order to circumvent this in the future, we now do the following:

test if we even need to drop DB triggers with a simple SQL query that does not need SUPER privileges
if the query can not be executed log a warning and try to drop the triggers
if dropping the triggers throws an exception log an appropriate error message and stop the update process