do not activate tab if not visible
Actually we want to avoid that the mobile version of the page will
break. If we call the click handler of the tab in the mobile version,
the location get's changed and an "#" is added. This may prevent the
action that should actually have happened, for example changing the
product variant, because the location change prevents the event
propagation - why ever - but it happens for stilwahl.de

So to avoid this, we just do not fire the click event for the tabs on
the mobile version, i.e. when the tab is not visible.