do not disconnect from CSE anymore
in some scenarios a shop just disconnected from CSE because there
was a deployment going on and plugins got disabled and enabled again.
to avoid such scenarios we now "lock" a tenant connection to the CSE
and only allow to disconnect the shop manually or via data-reset.

if the plugin is disabled/deinstalled the only thing that will happen
to the shop while it is not disconnected is just a HTTP request
to pull products that will go into nirvana because the plugin is not
installed anymore, i.e. the endpoint the product export is targeting is
not there.